+++ src/main/java/io/github/epi155/emsql/commons/SqlEnum.java
@@ -161,7 +161,7 @@ public enum SqlEnum implements SqlDataType {
     DecimalNil("BigDecimal", "NUMERIC") {
         @Override
         public Collection<String> requires() {
-            return NumberStd.requires();
+            return DecimalStd.requires();
         }
     },
     BinaryStd("Bytes", "BINARY", "byte[]"),