+++ src/main/java/io/github/epi155/emsql/spi/SqlParser.java
@@ -1,0 +1,7 @@
+package io.github.epi155.emsql.spi;
+
+import io.github.epi155.emsql.commons.SqlAction;
+
+public interface SqlParser {
+    void  validate(String query, Class<? extends SqlAction> action);
+}